backend/requirements.txt
Django>=4.2,<6.0
djangorestframework>=3.14,<4.0
djangorestframework-simplejwt>=5.3,<6.0
drf-spectacular>=0.27,<1.0
django-cors-headers>=4.3,<5.0
# Optional (only if you switch to Postgres later):
# psycopg2-binary>=2.9,<3.0
